Adding support for buffer list view context.
This allows a buffer to be updated using code like:

    with buf.list() as v:
        # Update the list 'v' using full range of Python list
        # methods and slicing.

Within the with-block, the list 'v' represents the buffer contents.
When the context is exited, the buffer is updated with the contents
of 'v'. An exception within the block leaves the buffer unchanged.

For non-trivial updates to a buffer, this form of code is much faster that
directly indexing the buffer object.