ScratchBuffer: display hook and namespaces.
Use autocmd to trigger on_first_showing call.  This is a cleaner mechanism that
counting number of entries into a ScratchBuffer.

A ScratchBuffer now has auto_grp_name and syntax_prefix properties.
These are generated from the new simple_name attribute and are thus
fairly unlikey to clash with prefixes.